requirements-prod.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# Text Extraction
PyPDF2==3.0.1
python-docx==1.1.2

# Async
aiofiles==24.1.0
httpx==0.28.1

# Web Search (DuckDuckGo fallback — no API key needed)
ddgs>=9.0.0


# Embeddings & ML
sentence-transformers>=3.0.0
numpy>=1.26.0

# Authentication
PyJWT==2.9.0

# Database (Azure SQL)
pyodbc==5.2.0

# Logging
structlog==24.4.0

# Runtime
python-dotenv==1.0.1
